Use cli_abort for missing resources in read_package(), but no longer require a name
The previous assertion checked for:

1. Resources property missing
2. Resources property = empty list
3. Resources do not have a name

That last assertion is removed, so it is now possible to read a package where not all the resources have a name. For any later step however, check_package() requires all resources to have a name. See class "frictionless_error_resources_without_name". I did not want to include this test in two places. Note also that name might not be required anymore in frictionless v2.

Ping @PietrH