Drop as.list(), use unclass() instead
Rather than making an as.list(), we mainly want people to be able to bypass the print method for datapackage. That can be done with unclass(), which will remove all classes and revert to the list primitive. This is more robust and similar to `class(package) <- "list" as was done in as.list()